refactor: Some structure refactor
