+++ src-rust/toolkit-cli/src/workflow/mod.rs
@@ -1,3 +1,1 @@
-mod select;
-
-pub use select::select;
+pub mod select;
