+++ src-rust/toolkit-core/Cargo.toml
@@ -22,7 +22,8 @@ regex = "1.10.5" # Regular Expression
 reqwest = { version = "0.12", features = ["json"] } # HTTP Client
 octocrab = "0.38.0" # GitHub API
 html2md = { workspace = true } # HTML to Markdown
-libhtmlfilter = "0.1.0" # HTML Filter
+libhtmlfilter = "0.1.2" # HTML Filter
+openai_api_rust = "0.1.9" # OpenAI API
 
 [dev-dependencies]
 wiremock = "0.6.0"
