+++ src-rust/toolkit-core/src/apis/mod.rs
@@ -1,3 +1,1 @@
-mod github_api;
-
-pub use github_api::*;
+pub mod github_api;
