+++ src-rust/toolkit-core/src/parser/mod.rs
@@ -1,3 +1,1 @@
-mod html_to_markdown;
-
-pub use html_to_markdown::html_to_markdown;
+pub mod html_to_markdown;
